Fix signed unsigned comparison

+++ include/mexplus/arguments.h
@@ -138,7 +138,7 @@ class InputArguments {
     for (entry = definitions_.begin(); entry != definitions_.end(); ++entry)
       if (!parseDefinition(nrhs, prhs, &entry->second))
         delete_positions.push_back(entry);
-    for (int i = 0; i < delete_positions.size(); ++i)
+    for (size_t i = 0; i < delete_positions.size(); ++i)
       definitions_.erase(delete_positions[i]);
     if (definitions_.empty())
       mexErrMsgIdAndTxt("mexplus:arguments:error",
@@ -215,7 +215,7 @@ class InputArguments {
   }
   /** Try to parse one definition or return false on failure.
    */
-  bool parseDefinition(int nrhs,
+  bool parseDefinition(size_t nrhs,
                        const mxArray* prhs[],
                        Definition* definition) {
     const size_t kMaxOptionNameSize = 64;
@@ -227,7 +227,7 @@ class InputArguments {
       error_message_.assign(message.str());
       return false;
     }
-    int index = 0;
+    size_t index = 0;
     for (; index < definition->mandatories.size(); ++index)
       definition->mandatories[index] = prhs[index];
 
@@ -412,7 +412,7 @@ class OutputArguments {
  private:
   /** Number of output arguments.
    */
-  int nlhs_;
+  size_t nlhs_;
   /** Output argument array.
    */
   mxArray** plhs_;
